+++ src/sonictune/ui/qml/pages/StatsPage.qml
@@ -10,6 +10,7 @@ import "../components"
 Item {
     id: statsPage
 
+    objectName: \"stats"
     property var stats: ({})
     property bool loading: true
     property bool loadError: false